+++ src/components/messaging/modal-dialog-actions/modal-dialog-actions.vue
@@ -1,0 +1,5 @@
+<template>
+	<div class="mt-12 flex flex-wrap items-center gap-6 border-t border-grey-200 dark:border-white/20 pt-6" data-test="modal-dialog-actions">
+		<slot />
+	</div>
+</template>